moved log file into a folder
