+++ loop/urls.py
@@ -8,7 +8,7 @@ from api import FarmerResource, VillageResource, LoopUserResource, CropResource,
 
 from loop.views import *
 
-from loop_data_log import send_updated_log
+from loop.utils.send_log.loop_data_log import send_updated_log
 
 api = Api(api_name = "v1")
 api.register(VillageResource())
